+++ CHANGES.md
@@ -3,6 +3,7 @@ Changes coming in 0.3.0-alpha6 (or beta1)
 * query as-arrays? now allows you to leverage lazy result fetching [JDBC-72](http://dev.clojure.org/jira/browse/JDBC-72).
 * "h2" is recognized as a protocol shorthand for org.h2.Driver
 * Tests no longer use :1 literal [JDBC-71](http://dev.clojure.org/jira/browse/JDBC-71).
+* Conditional use of javax.naming.InitialContext so it can be compiled on Android [JDBC-69](http://dev.clojure.org/jira/browse/JDBC-69).
 
 Changes in 0.3.0-alpha5